Feat : 검색필드 포커스 시 검색화면 렌더링
