+++ .env
@@ -1,0 +1,1 @@
+REACT_APP_SERVER_IP=http://localhost:4000/
